+++ packages/jellyfish-transaction/__tests__/tx_composer/index.test.ts
@@ -1,0 +1,29 @@
+import { SmartBuffer } from 'smart-buffer'
+import { ComposableBuffer } from '../../src/buffer/buffer_composer'
+
+// Test vector mostly taken from: https://github.com/bitcoin/bips/blob/master/bip-0143.mediawiki
+
+export function expectHexBufferToObject<T> (hex: string, data: T, asC: ((buffer: SmartBuffer) => ComposableBuffer<T>)): void {
+  const composable = asC(SmartBuffer.fromBuffer(
+    Buffer.from(hex, 'hex')
+  ))
+
+  expect(composable.toObject()).toEqual(data)
+  // parse and stringify due to JSON path inconsistent positioning
+  expect(JSON.parse(JSON.stringify(composable.toObject()))).toEqual(
+    JSON.parse(JSON.stringify(data))
+  )
+}
+
+export function expectObjectToHexBuffer<T> (data: T, hex: string, asC: ((data: T) => ComposableBuffer<T>)): void {
+  const txn = asC(data)
+
+  const buffer = new SmartBuffer()
+  txn.toBuffer(buffer)
+
+  expect(buffer.toBuffer().toString('hex')).toBe(hex)
+  // parse and stringify due to JSON path inconsistent positioning
+  expect(JSON.parse(JSON.stringify(txn.toObject()))).toEqual(
+    JSON.parse(JSON.stringify(data))
+  )
+}
